# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: NepalLend API
  version: 0.1.0
  description: |
    Contract for the NepalLend P2P lending backend.

    Conventions:
    - All monetary amounts are **integer paisa** (NPR × 100), field names end
      in `Paisa`. Never send or return floats for money.
    - Auth: `Authorization: Bearer <accessToken>` on everything except
      `/health` and `/auth/*`.
    - Errors: RFC 9457 problem shape `{ "error": "...", "message": "..." }`.
    - Lists accept `limit` / `offset` query params and return plain arrays.
    - Dates: RFC 3339 strings.
servers:
  - url: http://localhost:3000
    description: Local Express server
  - url: http://localhost:4001
    description: Prism mock
...
paths:
...
  /api/v1/auth/otp/request:
    post:
      tags: [auth]
      operationId: requestOtp
      summary: Send a login/registration OTP by SMS
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OtpRequest" }
      responses:
        "200":
          description: OTP sent (in dev/mock the code is `123456`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OtpChallenge" }
        "400":
          $ref: "#/components/responses/Unprocessable"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/auth/otp/verify:
    post:
      tags: [auth]
      operationId: verifyOtp
      summary: Verify the OTP and get session tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OtpVerify" }
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400":
          $ref: "#/components/responses/Unprocessable"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Revoke all refresh tokens for the current user
      security:
        - bearerAuth: []
      responses:
        "204": { description: Logged out }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Missing/invalid credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    TooManyRequests:
      description: Rate limited
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unprocessable:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string, examples: [validation_error] }
        message: { type: string }
...
    OtpRequest:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          pattern: "^9[0-9]{9}$"
          description: Nepali mobile number, 10 digits starting with 9
          examples: ["9841234567"]
    OtpChallenge:
      type: object
      required: [requestId, expiresInSeconds, phone]
      properties:
        requestId: { type: string, format: uuid }
        phone: { type: string, description: Masked number for display }
        maskedPhone: { type: string, examples: ["9841***567"] }
        expiresInSeconds: { type: integer, examples: [120] }
    OtpVerify:
      type: object
      required: [requestId, code]
      properties:
        requestId: { type: string, format: uuid }
        code: { type: string, minLength: 4, maxLength: 6, examples: ["123456"] }
    AuthSession:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresInSeconds: { type: integer, examples: [3600] }
        isNewUser: { type: boolean }
        user: { $ref: "#/components/schemas/User" }
    User:
      type: object
      required: [id, phone, role, createdAt]
      properties:
        id: { type: string, format: uuid }
        phone: { type: string, examples: ["9841234567"] }
        name: { type: [string, "null"], examples: ["Aarav Sharma"] }
        role: { type: string, enum: [borrower, lender] }
        email: { type: [string, "null"], format: email }
        avatarUrl: { type: [string, "null"], format: uri }
        kycState: { $ref: "#/components/schemas/KycState" }
        createdAt: { type: string, format: date-time }
...
    KycState:
      type: string
      enum: [unstarted, in_review, verified, rejected]
...